store the fixtures in data
