+++ tests/conftest.py
@@ -1,3 +1,2 @@
 """Configs for pytest."""
-from tests.fixtures.agent import *
-from tests.fixtures.retriever import *
+from tests.fixtures.data import *
