+++ tests/fixtures/data.py
@@ -11,6 +11,16 @@ from discussion_agents.cog.agent.react import ReActAgent
 from discussion_agents.cog.agent.reflexion import ReflexionCoTAgent, ReflexionReActAgent
 from pathlib import Path
 
+import faiss
+
+from langchain.docstore import InMemoryDocstore
+from langchain.embeddings import HuggingFaceEmbeddings
+from langchain.retrievers import TimeWeightedVectorStoreRetriever
+from langchain.vectorstores import FAISS
+
+
+
+
 @pytest.fixture
 def generative_agent() -> GenerativeAgent:
     """Creates a GenerativeAgent."""
@@ -61,3 +71,29 @@ def alfworld_env(alfworld_file):
     with open(alfworld_file) as reader:
         config = yaml.safe_load(reader) 
     return config
+
+"""Fixtures for creating retrievers."""
+
+
+embedding_size = (
+    768  # Embedding dimension for all-mpnet-base-v2. FAISS needs the same count.
+)
+model_name = "sentence-transformers/all-mpnet-base-v2"
+model_kwargs = {"device": "cpu"}
+encode_kwargs = {"normalize_embeddings": False}
+
+
+@pytest.fixture
+def time_weighted_retriever() -> TimeWeightedVectorStoreRetriever:
+    """Creates a TimeWeightedVectorStoreRetriever."""
+    embeddings_model = HuggingFaceEmbeddings(
+        model_name=model_name, model_kwargs=model_kwargs, encode_kwargs=encode_kwargs
+    )
+    index = faiss.IndexFlatL2(embedding_size)
+    vectorstore = FAISS(embeddings_model.embed_query, index, InMemoryDocstore({}), {})
+    retriever = TimeWeightedVectorStoreRetriever(
+        vectorstore=vectorstore, otherScoreKeys=["importance"], k=5
+    )
+    return retriever
+
+
